changed logic of change_location method
